telegram-vpn-bot/requirements.txt
aiogram~=3.5.0
aiohttp~=3.9.5
certifi==2024.2.2
ujson==5.9.0
environs==11.0.0
requests~=2.31.0
cachetools~=5.3.3
marzban-api-client~=0.4.9
betterlogging~=1.0.0
yookassa>=3.0.1
httpx>=0.27.0
aiosqlite>=0.20.0